+++ dmr/controller.py
@@ -1,6 +1,6 @@
 from collections.abc import Callable, Mapping, Sequence, Set
 from http import HTTPMethod, HTTPStatus
-from typing import Any, ClassVar, Final, Generic, TypeAlias, TypeVar
+from typing import TYPE_CHECKING, Any, ClassVar, Final, Generic, TypeAlias, TypeVar
 
 from django.http import HttpRequest, HttpResponse, HttpResponseBase
 from django.urls import URLPattern
@@ -27,6 +27,9 @@ from dmr.serializer import BaseSerializer
 from dmr.settings import HttpSpec
 from dmr.types import AnnotationsContext, infer_type_args
 from dmr.validation import ControllerValidator, SettingsValidator
+
+if TYPE_CHECKING:
+    from dmr.routing import Router
 
 _METHOD_NOT_ALLOWED_MSG: Final = _(
     'Method {method} is not allowed, allowed: {allowed}',
@@ -487,6 +490,7 @@ class Controller(Generic[_SerializerT_co], View):  # noqa: WPS214
         path: str,
         pattern: URLPattern,
         context: OpenAPIContext,
+        router: 'Router | None' = None,
     ) -> PathItem:
         """Generate OpenAPI spec for path items."""
         operations: dict[str, Any] = {
@@ -496,6 +500,7 @@ class Controller(Generic[_SerializerT_co], View):  # noqa: WPS214
                 cls.__qualname__,
                 cls.serializer,
                 context,
+                router=router,
             )
             for method, endpoint in cls.api_endpoints.items()
         }