+++ dmr/endpoint.py
@@ -47,6 +47,7 @@ from dmr.validation import (
 if TYPE_CHECKING:
     from dmr.controller import Controller
     from dmr.openapi.core.context import OpenAPIContext
+    from dmr.routing import Router
     from dmr.validation.response import ValidatedModification
 
 
@@ -262,6 +263,7 @@ class Endpoint:  # noqa: WPS214
         controller_name: str,
         serializer: type[BaseSerializer],
         context: 'OpenAPIContext',
+        router: 'Router | None' = None,
     ) -> Operation:
         """Build an OpenAPI Operation from an endpoint."""
         operation_id = self.get_operation_id(
@@ -281,11 +283,15 @@ class Endpoint:  # noqa: WPS214
             serializer,
         )
 
+        tags = list(router.tags) if router and router.tags else []
+        if self.metadata.tags:
+            tags.extend(self.metadata.tags)
+
         return Operation(
-            tags=self.metadata.tags,
+            tags=tags or None,
             summary=self.metadata.summary,
             description=self.metadata.description,
-            deprecated=self.metadata.deprecated,
+            deprecated=self.metadata.deprecated or (router and router.deprecated),
             security=security,
             external_docs=self.metadata.external_docs,
             servers=self.metadata.servers,